+++ verik-importer/src/main/kotlin/io/verik/importer/main/VerikImporterMain.kt
@@ -38,7 +38,7 @@ object VerikImporterMain {
         try {
             val projectContext = ProjectContext(config)
             readFiles(projectContext)
-            stageSequence.process(projectContext)
+            stageSequence.processAll(projectContext)
             writeFiles(projectContext)
         } catch (exception: Exception) {
             if (exception !is VerikImporterException) {