+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 ### Added
+- Support conditional compiler directives in importer.
+
+### Changed
+- Adopt `StageType` based stage sequence in importer.
 
 ## [0.1.11]
 ### Added